yazar silme koşulu eklendi.

+++ BookStoreApi/Applications/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -10,9 +10,13 @@ public class DeleteAuthorCommand
 	public void Handle()
 	{
 		var author = _dbContext.Authors.SingleOrDefault(a => a.Id == AuthorId);
+		var authorBooks = _dbContext.Books.SingleOrDefault(a => a.AuthorId == AuthorId);
 		
 		if (author is null)
 			throw new InvalidOperationException("ID isn't found.");
+			
+		if (authorBooks is not null)
+			throw new InvalidOperationException("The Author has published book. Please delete book first.");
 			
 		_dbContext.Authors.Remove(author);
 		_dbContext.SaveChanges();
